% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{un_roll_call_issues}
\alias{un_roll_call_issues}
\title{Data on the issue of each roll call of the United Nations General Assembly}
\format{A data frame (specifically a tbl_df) with one row
for each pair of a roll call vote and an issue describing that vote.
Contains the following columns:
\describe{
  \item{rcid}{The roll call id; used to join with \code{\link{un_votes}}
  and \code{\link{un_roll_calls}}}
  \item{short_name}{Two-letter issue codes}
  \item{issue}{Descriptive issue name}
}}
\source{
Erik Voeten "Data and Analyses of Voting in the UN General Assembly"
 Routledge Handbook of International Organization, edited by Bob Reinalda
  (published May 27, 2013)
  \url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=hdl:1902.1/12379}
}
\usage{
un_roll_call_issues
}
\description{
Issue (topic) classifications of roll call votes of the United Nations
General Assembly. Many votes had no topic, and some have more than one.
}
\keyword{datasets}
